+++ src/test/java/com/google/cloud/spanner/pgadapter/statements/StatementTest.java
@@ -63,8 +63,6 @@ import com.google.common.primitives.Bytes;
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
-import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -72,7 +70,6 @@ import java.util.Arrays;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import org.junit.AfterClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -103,19 +100,6 @@ public class StatementTest {
   @Mock private StatementResult statementResult;
   @Mock private ResultSet resultSet;
   @Mock private DataOutputStream outputStream;
-
-  @AfterClass
-  public static void cleanup() {
-    deleteLogFile();
-  }
-
-  private static void deleteLogFile() {
-    // TODO: Make error log file configurable and turn off writing to a file during tests.
-    try {
-      Files.deleteIfExists(new File("output.txt").toPath());
-    } catch (IOException ignore) {
-    }
-  }
 
   @Test
   public void testBasicSelectStatement() throws Exception {
@@ -676,7 +660,6 @@ public class StatementTest {
               return mw;
             });
     executor.shutdown();
-    ;
 
     backendConnection.flush();
 
@@ -687,8 +670,6 @@ public class StatementTest {
         thrown.getMessage());
 
     copyStatement.close();
-
-    deleteLogFile();
   }
 
   private void setupQueryInformationSchemaResults() {